# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short human-readable message
        type: string
    type: object
...
  internal_features_users_transport_http.GetUserResponse:
    properties:
      email:
        example: ivanivanov@example.ex
        type: string
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.PatchUserRequest:
    properties:
      email:
        example: jebron.lames@goat.forever
        type: string
      full_name:
        example: Jebron Lames
        type: string
      phone_number:
        example: "+71112223344"
        type: string
    type: object
  internal_features_users_transport_http.PatchUserResponse:
    properties:
      email:
        example: ivanivanov@example.ex
        type: string
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 3
        type: integer
    type: object
host: 127.0.0.1:5050
info:
  contact: {}
  description: Todo Application REST-API scheme
  title: Golang Todo API
  version: "1.0"
paths:
...
  /users/{id}:
    delete:
      description: Delete an existing user by id
      parameters:
      - description: ID of the user to be deleted
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Successful user deleting
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      description: Get a specific user by ID
      parameters:
      - description: ID of the received user
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User seccessfuly found
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.GetUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: |-
        Change existing user information
        ### Logic update fields (Three-state logic):
        1. **The field is not transmitted**: `phone_number` ignored, the value in the database does not change
        2. **Passed value**: `"phone_number`": "+71112223344"` - set a new phone number in the database
        3. **Passed null**: `"phone_number`": "null"` - clear a field in the database (set to NULL)
        Restrictions: `full_name ` can't be null
      parameters:
      - description: User id to change
        in: path
        name: id
        required: true
        type: integer
      - description: PatchUser request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Seccessfull changed user
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {access_token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"